requirements.txt
asgiref==3.2.7
awesome-slugify==1.6.5
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-heroku==0.3.1
gunicorn==20.0.4
numpy==1.18.4
pandas==1.0.4
patsy==0.5.1
Pillow==7.1.2
plotly==4.8.1
plotly-express==0.4.1
psycopg2==2.7.5
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
regex==2020.5.14
retrying==1.3.3
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
statsmodels==0.11.1
Unidecode==0.4.21
whitenoise==5.1.0
